+++ src/idlib/Str.h
@@ -404,7 +404,7 @@ ID_INLINE idStr::idStr(const char *text)
 	Init();
 	if (text)
 	{
-		l = strlen(text);
+		l = (int)strlen(text);
 		EnsureAlloced(l + 1);
 		strcpy(data, text);
 		len = l;
@@ -414,7 +414,7 @@ ID_INLINE idStr::idStr(const char *text)
 ID_INLINE idStr::idStr(const char *text, int start, int end)
 {
 	int i;
-	int l = strlen(text);
+	int l = (int)strlen(text);
 
 	Init();
 	if (end > l)
@@ -722,7 +722,7 @@ ID_INLINE int idStr::Cmpn(const char *text, int n) const
 ID_INLINE int idStr::CmpPrefix(const char *text) const
 {
 	assert(text);
-	return idStr::Cmpn(data, text, strlen(text));
+	return idStr::Cmpn(data, text, (int)strlen(text));
 }
 
 ID_INLINE int idStr::Icmp(const char *text) const
@@ -740,7 +740,7 @@ ID_INLINE int idStr::Icmpn(const char *text, int n) const
 ID_INLINE int idStr::IcmpPrefix(const char *text) const
 {
 	assert(text);
-	return idStr::Icmpn(data, text, strlen(text));
+	return idStr::Icmpn(data, text, (int)strlen(text));
 }
 
 ID_INLINE int idStr::IcmpNoColor(const char *text) const
@@ -764,7 +764,7 @@ ID_INLINE int idStr::IcmpnPath(const char *text, int n) const
 ID_INLINE int idStr::IcmpPrefixPath(const char *text) const
 {
 	assert(text);
-	return idStr::IcmpnPath(data, text, strlen(text));
+	return idStr::IcmpnPath(data, text, (int)strlen(text));
 }
 
 ID_INLINE int idStr::Length(void) const
@@ -832,7 +832,7 @@ ID_INLINE void idStr::Append(const char *text)
 
 	if (text)
 	{
-		newLen = len + strlen(text);
+		newLen = len + (int)strlen(text);
 		EnsureAlloced(newLen + 1);
 		for (i = 0; text[i]; i++)
 		{
@@ -897,7 +897,7 @@ ID_INLINE void idStr::Insert(const char *text, int index)
 		index = len;
 	}
 
-	l = strlen(text);
+	l = (int)strlen(text);
 	EnsureAlloced(len + l + 1);
 	for (i = len; i >= index; i--)
 	{
